Adding "smart" skip to moneyshot.
If the skip to money shot is not specified by you, then the app will try to figure out the most intense chunk of the script to skip to. (Note this is far from perfect but good enough I think. Some of the more "robotic scripts" will not work so well)